# backend/src/routes/mc.ts
import type { Request, Response } from 'express';
import { Router } from 'express';
import { serverAuthMiddleware } from '../middleware/server-auth.js';
import { ForbiddenError } from '../utils/errors.js';
import { validate, parseId, parsePagination } from '../utils/validate.js';
import * as authService from '../services/auth.service.js';
import * as mcService from '../services/mc.service.js';
import {
  mcCommentSchema,
  mcLinkCodeSchema,
  mcRegisterSchema,
  mcStatusSchema,
  mcTicketActionSchema,
  mcTicketSchema,
  mcUnlinkSchema,
  mcViewerSchema,
} from '../schemas/mc.js';

const router = Router();

router.use(serverAuthMiddleware);

router.post('/register', async (req: Request, res: Response) => {
  const data = validate(mcRegisterSchema, req.body);

  const { getSiteConfig } = await import('../services/setup.service.js');
  const siteConfig = await getSiteConfig();
  if (!siteConfig.allowMcRegister) {
    throw new ForbiddenError('Minecraft注册已关闭，请联系管理员');
  }

  const result = await authService.registerFromMinecraft(
    data.email,
    data.password,
    data.username,
    data.minecraftUuid,
    data.minecraftName,
  );
  res.status(201).json(result);
});

router.post('/link-code', async (req: Request, res: Response) => {
  const data = validate(mcLinkCodeSchema, req.body);

  const linkCode = await mcService.createLinkCode({
    minecraftUuid: data.minecraftUuid,
    minecraftName: data.minecraftName,
    serverId: req.server!.id,
  });

  res.status(201).json(linkCode);
});

router.post('/tickets', async (req: Request, res: Response) => {
  const data = validate(mcTicketSchema, req.body);

  const ticket = await mcService.createTicketFromMinecraft({
    minecraftUuid: data.minecraftUuid,
    title: data.title,
    body: data.body,
    template: data.template,
    formData: data.formData || {},
    serverId: req.server!.id,
    context: data.context,
    hidden: data.hidden,
  });

  res.status(201).json(ticket);
});

async function listMinecraftTickets(req: Request, res: Response, minecraftUuid?: string) {
  const { page, pageSize } = parsePagination(req.query as Record<string, unknown>);
  const result = await mcService.listTicketsForMinecraftViewer({
    page,
    pageSize,
    minecraftUuid,
  });
  res.json(result);
}

router.get('/tickets', async (req: Request, res: Response) => {
  const query = validate(mcViewerSchema, req.query);
  await listMinecraftTickets(req, res, query.minecraftUuid);
});

// Backward-compatible path for plugins from the previous release.
router.get('/tickets/:uuid', async (req: Request, res: Response) => {
  await listMinecraftTickets(req, res, String(req.params.uuid));
});

router.get('/tickets/:id/detail', async (req: Request, res: Response) => {
  const query = validate(mcViewerSchema, req.query);
  const ticket = await mcService.getTicketForMinecraft(
    parseId(String(req.params.id)),
    query.minecraftUuid,
  );
  res.json(ticket);
});

router.get('/tickets/:id/comments', async (req: Request, res: Response) => {
  const query = validate(mcViewerSchema, req.query);
  const comments = await mcService.listCommentsForMinecraft(
    parseId(String(req.params.id)),
    query.minecraftUuid,
  );
  res.json(comments);
});

router.get('/user/:uuid', async (req: Request, res: Response) => {
  const user = await mcService.getLinkedUser(String(req.params.uuid));
  res.json(user);
});

router.post('/comments', async (req: Request, res: Response) => {
  const { minecraftUuid, ticketId, body } = validate(mcCommentSchema, req.body);

  const comment = await mcService.createCommentFromMinecraft({
    minecraftUuid,
    ticketId,
    body,
  });
  res.status(201).json(comment);
});

router.post('/tickets/:id/close', async (req: Request, res: Response) => {
  const { minecraftUuid } = validate(mcTicketActionSchema, req.body);

  const ticket = await mcService.closeTicketFromMinecraft(
    parseId(String(req.params.id)),
    minecraftUuid,
  );
  res.json(ticket);
});

router.post('/tickets/:id/reopen', async (req: Request, res: Response) => {
  const { minecraftUuid } = validate(mcTicketActionSchema, req.body);

  const ticket = await mcService.reopenTicketFromMinecraft(
    parseId(String(req.params.id)),
    minecraftUuid,
  );
  res.json(ticket);
});

router.post('/tickets/:id/status', async (req: Request, res: Response) => {
  const data = validate(mcStatusSchema, req.body);

  const ticket = await mcService.updateTicketStatusFromMinecraft(
    parseId(String(req.params.id)),
    data,
  );
  res.json(ticket);
});

router.post('/unlink', async (req: Request, res: Response) => {
  const { minecraftUuid } = validate(mcUnlinkSchema, req.body);

  const user = await mcService.unlinkMinecraftByUuid(minecraftUuid);
  res.json(user);
});

export default router;